Makes expanding book data fully part of the connector
