+++ bookwyrm/connectors/__init__.py
@@ -1,4 +1,4 @@
 ''' bring connectors into the namespace '''
 from .settings import CONNECTORS
-from .abstract_connector import ConnectorException
+from .abstract_connector import ConnectorException, load_connector
 from .abstract_connector import get_data, get_image
